cmake: Fixes inclusion of export header
This commit fixes the way that the export header generated by CMake's
`generate_export_header()` was being used in the project. Previously,
the export header was only included in the source files. On Linux, this
was not problematic. However, on Windows with the Visual Studio
compiler, this lead to compile time errors due to symbols being defined
with different linkage. In the source symbols were being declared as
*exported*, while in the public header file, their linkage wasn't
specified. Moreover, the CMake configuration lacked the necessary
compiler definitions to enable/disable usage of *export*/*import* function
attributes. In particular, the CMake configuration did not define
`${BASE_NAME}_STATIC_DEFINE` when building as a static library, or
`<library>_EXPORTS` when building as a shared library.

This commit corrects these errors. Now, the export header is installed
along with the public headers and the directory where it is created
during CMake configuration is added as a public transitive dependency on this
library component. In addition, `EFIKA_STATIC_DEFINE` is made a public
transitive dependency when built as a static library so that projects
linking against this library component will have the correct
compiler definitions to disable *export*/*import* function attributes.
Finally, `efika_core_EXPORTS` is made a public dependency of the public
library and an interface dependency of the sdk libray when built as
a shared library so that public functions will be decorated with the
*export* attribute during the build (sdk library), and when linked
against, those same functions will be decorated with the *import*
attribute (public library).

Since the export header (`efika/core/export.h`) must be included in the
public header `efika/core.h`, it is no longer necessary to explicitly
include it in source files, so all occurrences of it have been removed.
The rationale is that the macros defined therein are now part of the
public API and should be treated as such, i.e., accessed via the public
header, not a private header.