cmake: Overhauls component framework
The original component framework was very much adhoc. This commit
updates this Efika component to be compatible with CMake `find_package()`
components. Now when a consumer of the Efika project uses
`find_package()` to link against the Efika::efika target, they can
supply a list of components that are required to be installed. If any of
said components were not built and installed along with libEfika, the
`find_package()` call will fail. Currently, the components are organized
so that other components may be installed, but any which are not
required will not propagate transitive dependencies to the consuming
project; yet, any required will.